third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-item-end-after-get-set-expected.txt

Test that setting and getting grid-end and grid-after works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting -webkit-grid-end and -webkit-grid-after set through CSS
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') should be auto / span 1 first. Was auto / auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') should be span 1 first. Was auto.
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') should be auto / span 1 last. Was auto / auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') should be span 1 last. Was auto.
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') should be auto / span 2 first. Was auto / auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') should be span 2 first. Was auto.
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') should be auto / span 3 last. Was auto / auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') should be span 3 last. Was auto.

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-after') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row') is 'auto / auto'

Test getting and setting grid-end and grid-after through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') should be auto / span 1 first. Was auto / auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') should be span 1 first. Was auto.
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') should be auto / span 1 last. Was auto / auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
FAIL getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') should be span 1 last. Was auto.
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue

Test setting grid-start and grid-before to 'inherit' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue

Test setting grid-start and grid-before to 'initial' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-column') is gridColumnValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-start') is gridStartValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-end') is gridEndValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-row') is gridRowValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-before') is gridBeforeValue
PASS getComputedStyle(gridItem, '').getPropertyValue('-webkit-grid-after') is gridAfterValue

Test setting grid-end and grid-after back to 'auto' through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-end') is '18'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column') is 'auto / 18'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-after') is '66'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row') is 'auto / 66'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-end') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-after') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-row') is 'auto / auto'
PASS successfullyParsed is true

TEST COMPLETE
